requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
bcrypt==5.0.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
firebase_admin==7.3.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
proto-plus==1.27.1
protobuf==6.33.6
psycopg[binary]==3.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
